Created unpack_dir function

+++ python/unzip_files_in_directory.py
@@ -1,12 +1,24 @@
 import os
 import zipfile
 from python.variables import lambda_folder
+import glob
 
-for item in os.listdir(lambda_folder):  # loop through items in dir
-    if item.endswith('.zip'):  # check for ".zip" extension
-        arch_name = lambda_folder + '/' + item  # get full path of files
-        print(arch_name)
-        zip_ref = zipfile.ZipFile(arch_name)  # create zipfile object
-        zip_ref.extractall(lambda_folder)  # extract file to dir
-        zip_ref.close()  # close file
-        os.remove(arch_name)  # delete zipped file
+
+def unpack_dir(mypath):
+    zipcounter = len(glob.glob1(mypath,"*.zip"))
+    if zipcounter == 0:
+        print("DONE!")
+    else:
+        for item in os.listdir(mypath):  # loop through items in dir
+            if item.endswith('.zip'):  # check for ".zip" extension
+                arch_name = mypath + '/' + item  # get full path of files
+                print(arch_name)
+                zip_ref = zipfile.ZipFile(arch_name)  # create zipfile object
+                zip_ref.extractall(mypath)  # extract file to dir
+                zip_ref.close()  # close file
+                os.remove(arch_name)  # delete zipped file
+        unpack_dir(mypath)
+
+
+unpack_dir(lambda_folder)
+
